backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
gunicorn==23.0.0
feedparser==6.0.11
beautifulsoup4==4.13.4
requests==2.32.3
python-dotenv==1.1.0
lxml==5.4.0
apscheduler==3.11.0
nltk==3.8.1
snownlp==0.12.3
scikit-learn==1.3.0
pandas==1.5.1
numpy==1.25.2
jieba==0.42.1
